backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
whisperx
pydub
rapidfuzz
google-generativeai
torch
aiofiles
